Implemented generate for the BayesClassifier

+++ c++_implementation/source/common.h
@@ -9,7 +9,7 @@
 //TODO proper types
 typedef boost::numeric::ublas::matrix<double>  	image_data_t;
 typedef std::vector<short> 						label_data_t;
-typedef boost::numeric::ublas::matrix<size_t>  	histogram_t;
+typedef boost::numeric::ublas::matrix<double>  	histogram_t;
 
 // stores the information for bins in given dimenstion
 struct bin_t
